Move all parser code over to the library
