+++ kicad_rs/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod resolver;
 pub mod types;
+pub mod parser;
